move test_no_query_adv_forum_search into test_search.py
also, correct xfail annotation and update README
