+++ tests/desktop/test_search.py
@@ -3,16 +3,32 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 from unittestzero import Assert
+from pages.desktop.search_page import SearchPage
 from pages.desktop.refine_search_page import RefineSearchPage
 import pytest
-xfail = pytest.mark.xfail
 
-class TestNoQueryAdvForumSearch:
+class TestSearch:
 
     @pytest.mark.smoketests
     @pytest.mark.bft
     @pytest.mark.fft
-    @xfail(reason='Bug 710361 - Empty/default advanced searches fail/time out')
+    @pytest.mark.prod
+    def test_cant_find_what_youre_looking_for_test(self, mozwebqa):
+        search_page_obj = SearchPage(mozwebqa)
+
+        searchTerms = ["firefox", "bgkhdsaghb"]
+        for current_search_term in searchTerms:
+            search_page_obj.go_to_search_page()
+            search_page_obj.do_search_on_search_box(current_search_term)
+
+            expected_text = "Can't find what you're looking for?"
+            Assert.contains(expected_text, search_page_obj.ask_a_question_text)
+            Assert.true(search_page_obj.is_ask_a_question_present, "Ask question link not present")
+
+    @pytest.mark.smoketests
+    @pytest.mark.bft
+    @pytest.mark.fft
+    @pytest.mark.xfail(reason='Bug 710361 - Empty/default advanced searches fail/time out')
     def test_no_query_adv_forum_search(self, mozwebqa):
         refine_search_pg = RefineSearchPage(mozwebqa)
